airflow/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.2
dbt-postgres==1.8.2
